+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for arrow keys, home, end, CTRL+arrow in line editor.
 - Added config file option to change shell used to execute commands.
 - Added sort options for interactive view. Defaults to sort by label.
+- Added command line switch to change command database file.
 
 ## [0.7.0] - 2021-02-09
 - Changed default search to search only labels, if no matches try searching commands.